pyproject.toml
[project]
name = "empirica-outreach"
version = "0.1.0"
description = "Epistemic marketing and community engagement platform"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Empirica Team"}
]
dependencies = [
    "empirica-ai>=1.1.1",  # Core dependency
    "praw>=7.7.0",         # Reddit API
    "click>=8.1.0",        # CLI framework
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
outreach = "empirica_outreach.cli.outreach_cli:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[tool.setuptools.packages.find]
include = ["empirica_outreach*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
